tokenizer-new: move tokenizeString and its helpers into module.f.ts
tokenizeString was test-only code in proof.f.ts, but it's the actual
tokenizer implementation (grammar -> AST -> flat tokens -> JsToken
list -> serialized string), so it and its transitive dependencies
(descentParserCpOnly, scanFunc, filterFunc, stringDecodeScan,
decodeJsonString, decodeNumber, toJsToken/toJsTokens,
getTokensFromAstRule and friends) belong in the module rather than
the test file. proof.f.ts now just imports tokenizeString and
descentParserCpOnly.

Aliased the ascii module's range (a [min,max] tuple builder) to
asciiRange in module.f.ts since it collided with the bnf module's
range (a TerminalRange encoder) already used throughout the grammar.
